Added seed-fetcher image to the build dir
the image depends on env vars to be able to retrive the seed file
CONJUR_SEED_FILE_URL - full url to get the seed file. e.g. 'https://conjur-master/configuration/seed/follower'
FOLLOWER_HOSTNAME - should be the name of the follower-service
SEEDFILE_DIR - the dir to save the seed file (should be mounted both to the follower and the seed-fetcher)