Use master service name when configuring master
The value used for the -h param when running evoke configure on mthe Master is
the value that shows up in the solo.json file as the Master's hostname. Using
the service name for this value allows containers from outside the namespace to
read the solo.json file and get a usable hostname for contacting a DAP Master.