Fix issues with follower deployments variable evaluation
Old code in some cases would get into configuration of the follower
without a CONJUR_DATA_KEY defined and bash would consider that to be an
error and return from the function without configuring the follower.
This change ensures that we go through the follower configuration even
if that variable is undefined (which is a common case in follower
deployment).