Split ServiceAccount & RoleBinding load
We would like to load only the RoleBinding in every namespace
that needs to use the authn-client. In this project we load
both resources and in external projects we will load only the RoleBinding.

Also, we renamed the service-account to `conjur-cluster` as this is
the service account that is used in the deployment.