Move common Summon vars into common section
Somewhere between Summon v0.6.6 and Summon v0.8.0 it became a requirement
that Summon vars **must** either be in a specific environment **or** in a `common`
section when using Summon environments.

This commit updates the secrets.yml for this project so that it conforms with this new
standard, to facilitate bumping the Summon version in our pipelines.

It also updates the Jenkinsfile to call the proper environment for the GKE tests.