Emphasize the meaning of STOP_RUNNING_ENV
The default value of STOP_RUNNING_ENV is set to 'true' so in case a customer
misses it, and has already created resources in their env, it will be deleted
without them noticing.

This commit adds a section to `bootstrap.env` that explains its meaning and also
echos that we are stopping the environment before doing so.