Adds an option to deploy without loadbalancer for Conjur master
This change adds an option to deploy without a LoadBalancer for Conjur
master. To deploy the Conjur service without a loadbalancer,
export the following:
      export DEPLOY_CONJUR_LOADBALANCER=false
The default value for this environment variable is `true`.

This change also removes a temporary workaround for this OpenShift 4.3
bug:
   https://bugzilla.redhat.com/show_bug.cgi?id=1798282
This workaround (i.e. deleting finalizers) can now be removed since
we can use a better workaround: Run OC 4.3 tests without a Conjur
loadbalancer to avoid the OC 4.3 bug.