Remove external facing load balancer
This removes the external facing load balancerfor the master cluster.
It was used exceedingly rarely and a bug in OpenShift 4.3 causes them
to not be cleaned up, in addition to requiring a special-case namespace
deletion to get it to even terminate.  A yaml for launching the
external facing load balancer is included for convenience, but not
deployed.

This removes special casing for OC 4.3 and enforces a timeout
on deleting the namespace. If a namespace fails to be deleted within
5 minutes it will result in the script exiting with an error.  The
previous special casing behavior for OC 4.3 would leave ELBs and
not properly clean up resources.