Eliminate 500 internal server errors when configuring CLI Pod
When the deploy scripts are run, when we get to the point where the Conjur
CLI Pod is being configured (in the `6.5_configure_cli_pod.sh` script),
we see the following error:

```
[2021-12-16T17:02:27.708Z] ++++++++++++++++++++++++++++++++++++++
[2021-12-16T17:02:27.708Z]
[2021-12-16T17:02:27.708Z] Configuring Conjur CLI.
[2021-12-16T17:02:27.708Z]
[2021-12-16T17:02:27.708Z] ++++++++++++++++++++++++++++++++++++++
     . . .
[2021-12-16T17:02:44.490Z] Logged in
     . . .
[2021-12-16T17:02:47.260Z] error: 500 Internal Server Error
     . . .
[2021-12-16T17:02:47.260Z] command terminated with exit code 1
```

This is an intermittent error that occurs when we're trying to log into
Conjur. I don't know the root cause, but executing Conjur CLI commands
inside a retry loop will fix the issue.